allow job-slurm.sql to run multiple times
job-slurm.sql currently errors out if run more than once; when using xdmod-setup to modify the db settings, etl_overseer is run with the xdmod.shredder-bootstrap.create-slurm-table action, but the 'table exists' error crashes the whole thing. 
Adding the 'IF NOT EXISTS' helps with this (and works for CentOS 7.6.1810, mysql Ver 15.1 Distrib 5.5.60-MariaDB).